backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.3.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
alembic==1.13.1
httpx==0.26.0
python-multipart==0.0.6

# AI & ML
langchain==0.3.22
langchain-anthropic==0.3.0
langchain-openai==0.3.12
langchain-community==0.3.20
langchain-core==0.3.49
langchain-text-splitters==0.3.7
# Keep <3 for compatibility with langchain==0.3.x / langchain-core<0.4
langchain-google-genai>=2.0.11,<3.0.0
anthropic>=0.16.0
scikit-learn==1.4.0
xgboost==2.0.3
pandas==2.1.4
numpy==1.26.3
joblib==1.3.2

# Security Scanners (CLI tools - install separately)
# snyk (npm install -g snyk)
# semgrep (pip install semgrep)

# Blockchain
web3==6.15.0
eth-account==0.10.0

# Database
psycopg2-binary==2.9.9  # PostgreSQL
aiosqlite==0.19.0       # SQLite async

# Utilities
pyyaml==6.0.1
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
